Rename CloudHistogram.vue -> BarChart.vue
